Add constants and build `ls` with the source
Added `ls` to necessary tables in `rs` and `kernel`.

Added constants relating to `ls`'s PID and available IPC message type
numbers to `com.h`.

Added a definition of the `ls` server to `/etc/system.conf`.

Added an entry to the `servers` Makefile to build `ls` alongside the
other servers.